`analog.dart` now uses HTTP server to recieve statements
`analog.dart` has an asyncronus HTTP server running which recieves statements sent from `scss.dart`.
When a statement is recieved, it is stored in a static list in the `StatementListener` class.
The statements in that list are then used to generate new Analog classes based off of the class names pulled from
the markup files.

There is currently a bug that prevents more than 1 type of a specific class generating.
I.E if you have a class named `p-1` and `p-2`, which ever class name comes first will be the one generated;
the other will be ignored.